iniciando desenvolvimento module, presentation e ajustando erros da feature details
