--- app/src/main/java/com/example/myapplication/home/presentation/base/BaseFragment.kt
@@ -1,29 +1,0 @@
-package com.example.myapplication.home.presentation.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.viewbinding.ViewBinding
-
-abstract class BaseFragment <VB: ViewBinding> : Fragment() {
-
-    private var _binding: VB? = null
-    protected val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = getViewBinding(inflater, container)
-        return _binding?.root
-    }
-
-    protected abstract fun getViewBinding(inflater: LayoutInflater, container: ViewGroup?): VB?
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
